@pixeltrix 2018-03-20
I think it'd be fine to use an inline ternary here to save adding an extra method, e.g:

``` ruby
"#{sanitized_method_name}#{multiple ? '[]' : ''}"
```

it's also shorter than the method call.